# Use sha1 from header if it exists