These headers would have been set by upstream in APKs. If they are present, just use them rather than creating temporary files and recomputing the sha1 for every file.

This cuts out ~25% of the work that `installAPKfiles` is doing and a couple seconds off the overall latency.

Before: 

```
apko publish --keyring-append  --repository-append  --arch amd64    13.74s user 5.02s system 215% cpu 8.721 total
```

<img width="997" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/a1120ecb-44ec-4b31-a6e0-e1a2ba9da2a4">

<img width="1165" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/8368b36f-32fc-4a41-938f-144f4f9e8281">

After:

```
apko publish --keyring-append  --repository-append  --arch amd64    12.61s user 3.45s system 238% cpu 6.746 total
```

<img width="995" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/6f6d4547-af06-43c9-9131-b1f552156056">

<img width="1159" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/7183ef7c-cbee-4d32-abc6-f1004fa62529">