@jonjohnsonjr 2023-07-08
With checksum confirmation:

```
apko publish --keyring-append  --repository-append  --arch amd64    13.45s user 3.81s system 225% cpu 7.655 total
```

<img width="882" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/3ea5a7d3-cbc5-4483-ab8c-4a6ccb019699">

<img width="1152" alt="image" src="https://github.com/chainguard-dev/go-apk/assets/17863526/9cbd61c2-f443-4a87-b732-868c7a769a85">